refactor(activities): the timeline's link rows become their own file
activity.go crossed the 500-line cap when the bound joined the dedupe at the
insert. A link is its own concept — a send, a booking, a logged note and a
captured message all reach it and share little else — so the vocabulary, the
bound, the refusals and the insert move together.
